Code Repository/requirements.txt
numpy==1.26.0
ipython==8.18.1
pandas==1.5.3
torch==2.0.1
torchvision==0.15.2  
torchaudio==2.0.2    
scipy==1.10.1
scikit-learn==1.2.2
seaborn==0.12.2
matplotlib==3.7.1
textstat==0.7.4
language-tool-python==2.8.1
spacy==3.5.3
nltk==3.8.1
tqdm==4.65.0
joblib==1.2.0
pyspellchecker==0.7.2
transformers==4.31.0
argparse==1.4.0
accelerate
peft
